# bump DEFAULT_CHUNK_THRESHOLD to 12