We now have that lovely SIMD by default so should probably think about that a bit more w.r.t the `Chunk` we choose.

Doing a benchmark such as

```jl
using ForwardDiff
using BenchmarkTools

import ForwardDiff.Dual
f(x) = (exp(x) * sin(x) + x^2 / tan(x))^3 / x^2
d_10 = Dual(rand(11)...)
d_12 = Dual(rand(13)...)

@btime f($d_10)
@btime f($d_12)
```

We find

```jl
julia> @btime f($d_10);
  77.918 ns (0 allocations: 0 bytes)

julia> @btime f($d_12);
  69.597 ns (0 allocations: 0 bytes)
```

so it is, in fact, faster to compute with chunk sizes of 12 instead of 10 (likely because they then all nicely fit into AVX registers and no need to clean up the scalars)

To see the effect on our favorite toy problem:

```jl
function rosenbrock(x)
   a = one(eltype(x))
   b = 100 * a
   result = zero(eltype(x))
   for i in 1:length(x)-1
       result += (a - x[i])^2 + b*(x[i+1] - x[i]^2)^2
   end
   return result
end

using BenchmarkTools
import ForwardDiff
const x = rand(60)
const y = similar(x)
const gradient_cache = ForwardDiff.GradientConfig(rosenbrock, x)
```

Before:

```jl
julia> @btime ForwardDiff.gradient!(y, rosenbrock, x, gradient_cache)
  5.589 μs (0 allocations: 0 bytes)
```

After:
```jl
julia> @btime ForwardDiff.gradient!(y, rosenbrock, x, gradient_cache)
  4.761 μs (0 allocations: 0 bytes)
```