@KristofferC 2018-11-05
Test failures due to some arbitrary chosen `FINITEDIFF_ERROR`. Yay...