@jrevels 2018-11-05
> Test failures due to some arbitrary chosen FINITEDIFF_ERROR. Yay...

Ugh, that's annoying. I really hate the finite differences approximation death cycle (want automated testing for AD -> need machine precision derivatives -> need AD -> want automated testing for AD...). More simple analytic unit tests would also be nice...

Bump the tolerance for those cases I guess? There are also methods for calculating finite difference approximation error bounds for a given function/input, it would be nice to incorporate those in the future (which ideally should be provided by the same package providing the finite diff implementation).