@KristofferC 2018-11-05
Bumped the tolerance a bit.